% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCART.R
\docType{package}
\name{QCART-package}
\alias{QCART-package}
\alias{QCART}
\title{R implementation of the QC-ART method}
\description{
This library consists of two functions: qcart and compute_threshold.  The former computes QC-ART scores given 
a dataset, baseline set and variables to use to compute the scores.  The latter uses the computed QC-ART scores 
to define threshold values that differentiate the nominal from outlying instrument runs.
}
