% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_scores.R
\name{qcart}
\alias{qcart}
\title{Computes QC-ART scores}
\usage{
qcart(all_data, baseline, variables, prop = 0.95)
}
\arguments{
\item{all_data}{a data frame of all of the data to be analyzed}

\item{baseline}{vector identifying which rows of `all_data` are the baseline observations?}

\item{variables}{vector of numbers or column names identifying which columns are the variables to use to compute scores}

\item{prop}{how many latent variables should be retained?  proportion of variability explianed by the latent variables}
}
\value{
vector of QC-ART scores corresponding to the rows of `all_data`
}
\description{
Computes QC-ART scores
}
