% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_funs.R
\name{compute_threshold}
\alias{compute_threshold}
\title{Computes threshold values for QC-ART Scores}
\usage{
compute_threshold(scores, baseline, time_stamps, type = "static",
  perc = 0.95)
}
\arguments{
\item{scores}{the QC-ART scores returned by `qcart`}

\item{baseline}{a vector of logical values (TRUE/FALSE) indicating which observations were used as the baseline}

\item{time_stamps}{the vector of time stamps assocaited with the QC-ART scores}

\item{type}{character specifying the type of threhsold to be returned: "static", "dynamic" or "both"}

\item{percentile}{what percentile of the model should be use as a cut-off, e.g. 0.9, 0.95, or 0.99}
}
\value{
a data.frame containing the QC-ART scores, baseline indicators, time stamps and threshold values for each qc-art score
}
\description{
This function will compute threshold values for QC-ART scores computed using the `qcart` function.
}
